Intro Scene, Dialogue System
Created intro scene that plays after the main scene.  Updated SceneMgr and Main Menu scene to route to Intro Scene instead of the Main Scene. DialogueAsset is a ScriptableObject that makes it easy to create dialogue systems in the game. Intro.cs includes dialogue helper functions that can be re-used.